Fix failing certificate renewal.
Proxying to the cerbot standalone web server was misconfigured in HAProxy.
The proxy destination was 'localhost', which is invalid inside the container.
Switched to ip-172-31-46-101.us-east-2.compute.internal like the CouchDB proxy uses.

The initial setup of the http01 challenge port was also broken.
Since HAProxy occupies port 80, we configured cerbot to start its server on a non-standard port and
proxied ACME challenge requests to it from HAProxy. Since HAProxy will fail to start if its certificate
file does not exist, certbot has to initially be launched in a standard configuration (standalone
webserver/http01_port on port 80) to acquire the first cert.

The certbot setup script assumed that 'http01_port =' would be present in the configuration after
this initial launch, but certbot omits it if the default (80) is used. Thus, the http01_port
replacement command did nothing and certbot was left to fight for port 80 with HAProxy.

http01_port is now added, if not already present, after the initial certbot run.